pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opengates-oss"
version = "0.1.0"
description = "Local-first, agent-native inbound gate runtime."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "jinja2>=3.1.4",
  "openai>=2.0.0",
  "pydantic>=2.9.0",
  "python-multipart>=0.0.12",
  "pyyaml>=6.0.2",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
]

[project.scripts]
opengates = "opengates.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
